Change IsBindingConfig() to an extension method
