+++ Celbridge/CoreServices/Celbridge.UserInterface/ViewModels/Forms/TextBoxElement.cs
@@ -122,7 +122,7 @@ public partial class TextBoxElement : FormElement
     {
         if (config.TryGetProperty("isEnabled", out var configValue))
         {
-            if (PropertyBinder<bool>.IsBindingConfig(configValue))
+            if (configValue.IsBindingConfig())
             {
                 _isEnabledBinder = PropertyBinder<bool>.Create(textBox, this)
                     .Setter((value) =>
@@ -217,7 +217,7 @@ public partial class TextBoxElement : FormElement
     {
         if (config.TryGetProperty("text", out var configValue))
         {
-            if (PropertyBinder<string>.IsBindingConfig(configValue))
+            if (configValue.IsBindingConfig())
             {
                 _textBinder = PropertyBinder<string>.Create(textBox, this)
                 .Binding(TextBox.TextProperty, BindingMode.TwoWay, nameof(Text))
